requirements.txt
fastapi
uvicorn[standard]
pytest
python-dotenv
huggingface_hub
sqlite-utils
